pvr: HACK: Allow -lsrv_{init,um}/-lIMGegl as LDLIBS
Use the LOCAL_ALLOWED_LDLIBS whitelisting introduced in build/make,
required for proper linking against the legacy libs without SONAME.

Ref: https://gerrit.unlegacy-android.org/#/c/3710/

NOTE:
I've changed the LOCAL_ADDITIONAL_DEPENDENCIES due to build error
of recepie not found.

However, the build still works since the libs are already installed
to $(TARGET_OUT_VENDOR).
This may not be the case on a *clean* build.

Change-Id: I7c983bc450d97961c1c7b52a9adac17ae7e6e558